Start dest driving the death of Location calls
