+++ corehq/apps/locations/dbaccessors.py
@@ -3,7 +3,7 @@ from corehq.apps.users.models import CommCareUser
 
 def get_users_by_location_id(location_id):
     """
-        Get all users for a given location
+    Get all users for a given location
     """
     return CommCareUser.view(
         'locations/users_by_location_id',
