+++ corehq/apps/locations/tests/__init__.py
@@ -4,5 +4,5 @@ from .test_location_set import *
 from .test_location_types import *
 from .test_locations import *
 from .test_products_at_location import *
-from .test_shim import *
+from .test_reupholster import *
 from .test_site_code import *
