--- corehq/apps/locations/tests/test_shim.py
@@ -1,16 +1,0 @@
-from .test_locations import LocationTestBase
-from .util import make_loc
-
-
-class TestPath(LocationTestBase):
-    def test_path(self):
-        locs = [
-            ('Mass', 'state'),
-            ('Suffolk', 'district'),
-            ('Boston', 'block'),
-        ]
-        parent = None
-        for name, type_ in locs:
-            parent = make_loc(name, type=type_, parent=parent)
-        boston = parent
-        self.assertEqual(boston.path, boston.sql_location.path)
